In-window menu impl no longer creates windows for the menus
This should fix the performance problem of menus that Windows and Linux
were exhibiting due to their insistence on doing animation during window
creation (at least when using glfw, which we are).

Note that I still need to go in and do size restrictions / better
placement due to this change.