Fix https://github.com/richardwilkes/gcs/issues/523
X11 responds asynchronously to window positioning and sizing requests.
Due to this, we need to wait for it to catch up, or subsequent code that
is relying on the coordinates being updated will get the wrong
information. We do put a cap on the amount of time we are willing to
wait, however, to ensure we don't hang should something go wrong.